split rcheevos.h into rc_runtime.h and rc_runtime_types.h
