+++ src/rcheevos/consoleinfo.c
@@ -1,4 +1,3 @@
-#include "rcheevos.h"
 #include "rc_consoles.h"
 
 #include <ctype.h>
