+++ src/rcheevos/runtime_progress.c
@@ -1,3 +1,4 @@
+#include "rc_runtime.h"
 #include "rc_internal.h"
 
 #include "../rhash/md5.h"
