+++ test/rcheevos-test.vcxproj
@@ -209,6 +209,8 @@
     <ClInclude Include="..\include\rc_api_user.h" />
     <ClInclude Include="..\include\rc_consoles.h" />
     <ClInclude Include="..\include\rc_hash.h" />
+    <ClInclude Include="..\include\rc_runtime.h" />
+    <ClInclude Include="..\include\rc_runtime_types.h" />
     <ClInclude Include="..\include\rc_url.h" />
     <ClInclude Include="..\src\rcheevos\rc_compat.h" />
     <ClInclude Include="..\src\rcheevos\rc_internal.h" />
