+++ test/rcheevos/test_consoleinfo.c
@@ -1,5 +1,4 @@
 #include "rc_consoles.h"
-#include "rcheevos.h"
 
 #include "../test_framework.h"
