+++ test/rcheevos/test_runtime.c
@@ -1,3 +1,4 @@
+#include "rc_runtime.h"
 #include "rc_internal.h"
 
 #include "mock_memory.h"
