+++ test/rcheevos/test_runtime_progress.c
@@ -1,3 +1,4 @@
+#include "rc_runtime.h"
 #include "rc_internal.h"
 
 #include "../test_framework.h"
